ESlint rule configured, messages + time showing
