+++ code/.eslintrc.json
@@ -68,6 +68,7 @@
     ],
     "no-param-reassign": "off",
     "no-restricted-syntax": "off",
+    "no-underscore-dangle": "off",
     "object-curly-newline": [
       "error",
       {
